Complete refactoring of the API.
This simplifies the use of our API, and allows us to define arbitrary types, such as for const factor division, in the future.
