--- src/ints/mod.rs
@@ -1,9 +1,0 @@
-//! Big integer definitions..
-
-#[macro_use]
-mod int_macros;
-#[macro_use]
-mod uint_macros;
-
-pub(crate) mod i256;
-pub(crate) mod u256;
